rename : components 위치 변경
