+++ src/app/(auth)/_components/CommonAgreeLayout.tsx
