+++ src/app/(auth)/_components/CommonLayout.tsx
