+++ src/app/(auth)/_components/Privacy.tsx
