+++ src/app/(auth)/_components/Terms.tsx
