+++ src/app/(auth)/register/page.tsx
@@ -1,5 +1,5 @@
-import CommonLayout from "@/components/auth/CommonLayout";
-import RegisterForm from "@/components/auth/RegisterForm";
+import CommonLayout from "../_components/CommonLayout";
+import RegisterForm from "../_components/RegisterForm";
 
 export default function RegisterPage() {
   return (
